+++ src/backend/plugins/video/warren_video/api.py
@@ -5,6 +5,7 @@ from warren_video.indicator.complete_viewers.api import (
     router as complete_viewers_router,
 )
 from warren_video.indicator.complete_views.api import router as complete_views_router
+from warren_video.indicator.downloads.api import router as downloads_router
 from warren_video.indicator.viewers.api import router as viewers_router
 from warren_video.indicator.views.api import router as views_router
 
@@ -14,3 +15,4 @@ router.include_router(views_router)
 router.include_router(complete_views_router)
 router.include_router(viewers_router)
 router.include_router(complete_viewers_router)
+router.include_router(downloads_router)